requirements.txt
Babel==2.9.1
mysql-connector==2.2.9
Pillow==9.0.1
pytz==2021.3
tkcalendar==1.6.1
bcrypt